requirements.txt
Flask
Werkzeug
Pillow
requests
web3
eth-account
eth-keys
eth-utils
subprocess